refactor: InitRoomPage 컴포넌트 분리 및 zustand 도입
