+++ client/src/store/index.tsx
@@ -1,0 +1,12 @@
+import create from 'zustand';
+import { NAVER_LAT, NAVER_LNG } from '@constants/map';
+
+interface MeetLocationStoreType {
+  meetLocation: { lat: number; lng: number };
+  updateMeetLocation: (lat: number, lng: number) => void;
+}
+
+export const useMeetLocationStore = create<MeetLocationStoreType>((set) => ({
+  meetLocation: { lat: NAVER_LAT, lng: NAVER_LNG },
+  updateMeetLocation: (lat, lng) => set(() => ({ meetLocation: { lat, lng } })),
+}));
